+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `--input-format` global option [#1038]
+
 ## [1.9.8] - 2025-05-15
 
 ### Added